blog_search_cloud/requirements.txt
# FastAPI 및 서버
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2

# HTTP 클라이언트
requests==2.32.3

# 환경 변수 관리
python-dotenv==1.0.1

# 검색 엔진 라이브러리
tavily-python==0.5.0
duckduckgo-search==6.3.5
